test: quantity: Split into multiple object files
The number of templates exceeded the symbol limit on
some platforms when built with debug symbols.
